fix(tui): address pet review feedback
