+++ codex-rs/tui/src/app_event.rs
@@ -338,6 +338,12 @@ pub(crate) enum AppEvent {
         result: Result<Option<crate::pets::AmbientPet>, String>,
     },
 
+    /// Result of restoring the configured ambient pet during startup.
+    ConfiguredPetLoaded {
+        pet_id: String,
+        result: Result<Option<crate::pets::AmbientPet>, String>,
+    },
+
     /// Refresh app connector state and mention bindings.
     RefreshConnectors {
         force_refetch: bool,
