+++ codex-rs/tui/src/chatwidget.rs
@@ -127,7 +127,6 @@ use codex_chatgpt::connectors as chatgpt_connectors;
 use codex_config::ConfigLayerStackOrdering;
 use codex_config::types::ApprovalsReviewer;
 use codex_config::types::Notifications;
-use codex_config::types::TuiPetAnchor;
 use codex_config::types::WindowsSandboxModeToml;
 use codex_core_skills::model::SkillMetadata;
 use codex_exec_server::EnvironmentManager;
@@ -339,6 +338,7 @@ use self::interrupts::InterruptManager;
 mod keymap_picker;
 mod mcp_startup;
 use self::mcp_startup::McpStartupStatus;
+mod pets;
 mod session_header;
 use self::session_header::SessionHeader;
 mod hooks;
@@ -4807,7 +4807,13 @@ impl ChatWidget {
             &chat_keymap.edit_queued_message,
             current_terminal_info,
         );
-        let ambient_pet = load_ambient_pet(&config, frame_requester.clone());
+        let ambient_pet = pets::load_ambient_pet(&config, frame_requester.clone());
+        pets::start_configured_pet_load_if_needed(
+            &config,
+            ambient_pet.is_none(),
+            frame_requester.clone(),
+            app_event_tx.clone(),
+        );
         let mut widget = Self {
             app_event_tx: app_event_tx.clone(),
             frame_requester: frame_requester.clone(),
@@ -6330,74 +6336,6 @@ impl ChatWidget {
         self.frame_requester.schedule_frame();
     }
 
-    fn set_ambient_pet_notification(
-        &mut self,
-        kind: crate::pets::PetNotificationKind,
-        body: Option<String>,
-    ) {
-        if let Some(pet) = self.ambient_pet.as_mut() {
-            pet.set_notification(kind, body);
-        }
-    }
-
-    pub(crate) fn ambient_pet_image_enabled(&self) -> bool {
-        self.ambient_pet
-            .as_ref()
-            .is_some_and(crate::pets::AmbientPet::image_enabled)
-    }
-
-    pub(crate) fn ambient_pet_draw(
-        &self,
-        area: Rect,
-        composer_bottom_y: u16,
-    ) -> Option<crate::pets::AmbientPetDraw> {
-        self.bottom_pane.no_modal_or_popup_active().then(|| {
-            let anchor_bottom_y = match self.config.tui_pet_anchor {
-                TuiPetAnchor::Composer => composer_bottom_y,
-                TuiPetAnchor::ScreenBottom => area.bottom(),
-            };
-            self.ambient_pet
-                .as_ref()?
-                .draw_request(area, anchor_bottom_y)
-        })?
-    }
-
-    fn ambient_pet_wrap_reserved_cols(&self) -> u16 {
-        self.ambient_pet
-            .as_ref()
-            .filter(|pet| pet.image_enabled())
-            .map(|pet| {
-                pet.image_columns()
-                    .saturating_add(AMBIENT_PET_WRAP_GAP_COLUMNS)
-            })
-            .unwrap_or(0)
-    }
-
-    pub(crate) fn history_wrap_width(&self, width: u16) -> u16 {
-        width
-            .saturating_sub(self.ambient_pet_wrap_reserved_cols())
-            .max(1)
-    }
-
-    pub(crate) fn pet_picker_preview_draw(
-        &self,
-        _terminal_area: Rect,
-    ) -> Option<crate::pets::AmbientPetDraw> {
-        self.bottom_pane
-            .selected_index_for_active_view(crate::pets::PET_PICKER_VIEW_ID)?;
-        let area = self.pet_picker_preview_state.area()?;
-        let request = self
-            .pet_picker_preview_pet
-            .as_ref()?
-            .preview_draw_request(area)?;
-        self.pet_picker_preview_image_visible.set(true);
-        Some(request)
-    }
-
-    pub(crate) fn should_clear_pet_picker_preview_image(&self) -> bool {
-        self.pet_picker_preview_image_visible.replace(false)
-    }
-
     fn bump_active_cell_revision(&mut self) {
         self.transcript.bump_active_cell_revision();
     }
@@ -6670,60 +6608,6 @@ impl ChatWidget {
             terminal_width,
         );
         self.bottom_pane.show_selection_view(params);
-    }
-
-    fn open_pets_picker(&mut self) {
-        if self.warn_if_pets_unsupported() {
-            return;
-        }
-
-        self.pet_picker_preview_state.clear();
-        self.pet_picker_preview_pet = None;
-        let params = crate::pets::build_pet_picker_params(
-            self.config.tui_pet.as_deref(),
-            &self.config.codex_home,
-            self.pet_picker_preview_state.clone(),
-        );
-        self.bottom_pane.show_selection_view(params);
-        let initial_pet_id = self
-            .config
-            .tui_pet
-            .clone()
-            .unwrap_or_else(|| crate::pets::DEFAULT_PET_ID.to_string());
-        self.start_pet_picker_preview(initial_pet_id);
-    }
-
-    fn select_pet_by_id(&mut self, pet_id: String) {
-        if self.warn_if_pets_unsupported() {
-            return;
-        }
-
-        self.app_event_tx.send(AppEvent::PetSelected { pet_id });
-    }
-
-    fn warn_if_pets_unsupported(&mut self) -> bool {
-        let support = self.pet_image_support();
-        let Some(message) = support.unsupported_message() else {
-            return false;
-        };
-
-        self.add_warning_message(message.to_string());
-        true
-    }
-
-    fn pet_image_support(&self) -> crate::pets::PetImageSupport {
-        #[cfg(test)]
-        if let Some(support) = self.pet_image_support_override {
-            return support;
-        }
-
-        #[cfg(test)]
-        return crate::pets::PetImageSupport::Unsupported(
-            crate::pets::PetImageUnsupportedReason::Terminal,
-        );
-
-        #[cfg(not(test))]
-        crate::pets::detect_pet_image_support()
     }
 
     fn status_line_context_window_size(&self) -> Option<i64> {
@@ -9049,145 +8933,6 @@ impl ChatWidget {
         self.config.tui_theme = theme;
     }
 
-    /// Set the pet preselected by the TUI picker in the widget's config copy.
-    pub(crate) fn set_tui_pet(&mut self, pet: Option<String>) {
-        self.config.tui_pet = pet;
-        self.ambient_pet = load_ambient_pet(&self.config, self.frame_requester.clone());
-        #[cfg(test)]
-        if let Some(support) = self.pet_image_support_override
-            && let Some(pet) = self.ambient_pet.as_mut()
-        {
-            pet.set_image_support_for_tests(support);
-        }
-        self.request_redraw();
-    }
-
-    pub(crate) fn set_tui_pet_loaded(
-        &mut self,
-        pet: Option<String>,
-        ambient_pet: Option<crate::pets::AmbientPet>,
-    ) {
-        self.config.tui_pet = pet;
-        self.ambient_pet = ambient_pet;
-        #[cfg(test)]
-        if let Some(support) = self.pet_image_support_override
-            && let Some(pet) = self.ambient_pet.as_mut()
-        {
-            pet.set_image_support_for_tests(support);
-        }
-        self.request_redraw();
-    }
-
-    pub(crate) fn start_pet_picker_preview(&mut self, pet_id: String) {
-        self.pet_picker_preview_request_id =
-            self.pet_picker_preview_request_id.wrapping_add(/*rhs*/ 1);
-        let request_id = self.pet_picker_preview_request_id;
-        self.pet_picker_preview_pet = None;
-        if pet_id == crate::pets::DISABLED_PET_ID {
-            self.pet_picker_preview_state.set_disabled();
-            self.request_redraw();
-            return;
-        }
-
-        self.pet_picker_preview_state.set_loading();
-        self.request_redraw();
-
-        let codex_home = self.config.codex_home.clone();
-        let frame_requester = self.frame_requester.clone();
-        let tx = self.app_event_tx.clone();
-        spawn_pet_load(move || {
-            let result = crate::pets::ensure_builtin_pack_for_pet(&pet_id, &codex_home)
-                .and_then(|()| {
-                    crate::pets::AmbientPet::load(
-                        Some(&pet_id),
-                        &codex_home,
-                        frame_requester,
-                        /*animations_enabled*/ false,
-                    )
-                })
-                .map_err(|err| err.to_string());
-            tx.send(AppEvent::PetPreviewLoaded { request_id, result });
-        });
-    }
-
-    pub(crate) fn finish_pet_picker_preview_load(
-        &mut self,
-        request_id: u64,
-        result: Result<crate::pets::AmbientPet, String>,
-    ) {
-        if request_id != self.pet_picker_preview_request_id {
-            return;
-        }
-
-        match result {
-            Ok(pet) => {
-                self.pet_picker_preview_state.set_ready();
-                self.pet_picker_preview_pet = Some(pet);
-                #[cfg(test)]
-                if let Some(support) = self.pet_image_support_override
-                    && let Some(pet) = self.pet_picker_preview_pet.as_mut()
-                {
-                    pet.set_image_support_for_tests(support);
-                }
-            }
-            Err(message) => {
-                self.pet_picker_preview_state.set_error(message);
-                self.pet_picker_preview_pet = None;
-            }
-        }
-        self.request_redraw();
-    }
-
-    pub(crate) fn show_pet_selection_loading_popup(&mut self) -> u64 {
-        self.pet_selection_load_request_id =
-            self.pet_selection_load_request_id.wrapping_add(/*rhs*/ 1);
-        self.pet_picker_preview_state.clear();
-        self.pet_picker_preview_pet = None;
-        self.bottom_pane.show_selection_view(SelectionViewParams {
-            view_id: Some(PET_SELECTION_LOADING_VIEW_ID),
-            title: Some("Loading Pet".to_string()),
-            subtitle: Some("Preparing the terminal pet.".to_string()),
-            items: vec![SelectionItem {
-                name: "Loading selected pet...".to_string(),
-                is_disabled: true,
-                ..Default::default()
-            }],
-            ..Default::default()
-        });
-        self.pet_selection_load_request_id
-    }
-
-    pub(crate) fn finish_pet_selection_loading_popup(&mut self, request_id: u64) -> bool {
-        if request_id != self.pet_selection_load_request_id {
-            return false;
-        }
-        self.bottom_pane
-            .dismiss_active_view_if_id(PET_SELECTION_LOADING_VIEW_ID);
-        true
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_pet_image_support_for_tests(
-        &mut self,
-        support: crate::pets::PetImageSupport,
-    ) {
-        self.pet_image_support_override = Some(support);
-        if let Some(pet) = self.ambient_pet.as_mut() {
-            pet.set_image_support_for_tests(support);
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn install_test_ambient_pet_for_tests(&mut self, animations_enabled: bool) {
-        self.set_tui_pet_loaded(
-            Some("test".to_string()),
-            Some(crate::pets::test_ambient_pet(
-                self.frame_requester.clone(),
-                animations_enabled,
-            )),
-        );
-    }
-
     /// Set the model in the widget's config copy and stored collaboration mode.
     pub(crate) fn set_model(&mut self, model: &str) {
         self.current_collaboration_mode = self.current_collaboration_mode.with_updates(
@@ -10849,32 +10594,6 @@ impl Drop for ChatWidget {
     fn drop(&mut self) {
         self.reset_realtime_conversation_state();
         self.stop_rate_limit_poller();
-    }
-}
-
-fn load_ambient_pet(
-    config: &Config,
-    frame_requester: FrameRequester,
-) -> Option<crate::pets::AmbientPet> {
-    let selected_pet = config.tui_pet.as_deref()?;
-    if selected_pet == crate::pets::DISABLED_PET_ID {
-        return None;
-    }
-
-    crate::pets::AmbientPet::load(
-        Some(selected_pet),
-        &config.codex_home,
-        frame_requester,
-        config.animations,
-    )
-    .ok()
-}
-
-fn spawn_pet_load(f: impl FnOnce() + Send + 'static) {
-    if let Ok(handle) = tokio::runtime::Handle::try_current() {
-        std::mem::drop(handle.spawn_blocking(f));
-    } else {
-        let _ = std::thread::spawn(f);
     }
 }
